Modifying java version 8 to 17, Dockerfile modification, .gitignore modification
