--- README.md
@@ -1,34 +1,0 @@
-# Getting Started with Spring, Maven and Docker
-
-## Prerequisite
-
-- JDK 18.0.1 or later 
-- Maven 3.8.5+
-- Spring Boot CLI(optional)
-- Microsoft Visual Studio Code
-- Docker Desktop
-
-
-## Clone the repo
-
-```
-git clone https://github.com/ajeetraina/spring-docker-demo
-```
-
-
-## Execute the below command to build the package:
-
-
-```
-mvn package
-```
-
-## Running the app packages as JAR file
-
-
-```
-java -jar target/spring-boot-docker-0.0.1-SNAPSHOT.jar 
-```
-
-You can now access the Hello World page via the web browser  http://localhost:8080 
-
