--- src/main/java/com/example/springbootdocker/DemoSpringbootApplication.java
@@ -1,21 +1,0 @@
-package com.example.springbootdocker;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@SpringBootApplication
-public class DemoSpringbootApplication {
-
-	@RequestMapping("/")
-	public String home() {
-		return "Hello World!";
-	}
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoSpringbootApplication.class, args);
-	}
-
-}
