--- src/test/java/com/example/springbootdocker/DemoSpringbootApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.springbootdocker;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoSpringbootApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
