+++ src/test/java/com/example/springbootdocker/SpringBootDockerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.springbootdocker;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBootDockerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
